PullRequest: 850 Remove logits dtype cast
Merge branch 'sglang_public_tracker_remove_logits_dtype_cast of git@code.alipay.com:Theta/SGLang.git into sglang_public_tracker

https://code.alipay.com/Theta/SGLang/pull_requests/850




* Remove logits dtype cast

+++ python/sglang/srt/models/bailing_moe_v3.py
@@ -448,9 +448,7 @@ class BailingMoEGate(nn.Module):
             self.expert_bias = None
 
     def forward(self, hidden_states):
-        logits = F.linear(hidden_states.to(self.weight.dtype), self.weight, None).to(
-            hidden_states.dtype
-        )
+        logits = F.linear(hidden_states.to(self.weight.dtype), self.weight, None)
         return logits
